Switch to symfony2 standards + concat with spaces.
In order to easily ensure that all developers are following the same coding standard, we are switching to php-cs-fixer. Symfony has some really nice additions to PSR-2.

Since PhpStorm doesn't support all the functionality of php-cs-fixer, it's harder to implement the same standards in both. Using Symfony allows for a more generic, and better set of standards.