Removed issue prefix from commit-standard.
Removed the issue prefix from the commit standards. The reasoning is that it will significantly reduce the amount of characters one has in the main commit message. Also with as many commits as one may have in a pull request, there is no reason to have every single commit tagged.